Add new features: Add API Gateway, Config server, and add some unit tests.

+++ microservice-product/src/main/java/com/microservice/product/infrastructure/dto/ProductResponseDTO.java
@@ -3,11 +3,12 @@ package com.microservice.product.infrastructure.dto;
 import lombok.Builder;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.UUID;
 
 /**
  * DTO for {@link com.microservice.product.domain.Product}
  */
 @Builder
-public record ProductResponseDTO(UUID id, String name, Double price) implements Serializable {
+public record ProductResponseDTO(UUID id, String name, BigDecimal price) implements Serializable {
 }
