+++ microservice-product/src/main/java/com/microservice/product/infrastructure/dto/SaveNewProductDTO.java
@@ -1,9 +1,10 @@
 package com.microservice.product.infrastructure.dto;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * DTO for {@link com.microservice.product.domain.Product}
  */
-public record SaveNewProductDTO(String name, Double price, Integer initialQuantity) implements Serializable {
+public record SaveNewProductDTO(String name, BigDecimal price, Integer initialQuantity) implements Serializable {
 }
